Move 'ensure submission' functionality to a service class
