+++ src/main/kotlin/com/terraformation/backend/accelerator/api/ParticipantProjectSpeciesController.kt
@@ -36,7 +36,7 @@ class ParticipantProjectSpeciesController(
   fun assignParticipantProjectSpecies(
       @RequestBody payload: AssignParticipantProjectSpeciesPayload
   ): SimpleSuccessResponsePayload {
-    participantProjectSpeciesStore.createMany(payload.projectIds, payload.speciesIds)
+    participantProjectSpeciesStore.create(payload.projectIds, payload.speciesIds)
     return SimpleSuccessResponsePayload()
   }
